+++ styles/globals.css
@@ -23,7 +23,3 @@
     flex-direction: column;
   }
 }
-
-.transition-all {
-  transition: all 0.3s ease;
-}
