+++ utils/helpers.ts
@@ -18,3 +18,10 @@ export const titleToAnchor = (title: string) => {
     .replace(/[\/\\^$*+?.()|\[\]{}<>:;"'~,=@`#!%&]/g, '')
     .toLowerCase()
 }
+
+export const truncateString = (str: string, maxLength: number) => {
+  if (str.length <= maxLength) {
+    return str
+  }
+  return str.slice(0, maxLength) + '...'
+}
